App in working state ready to go once documentation is finished

+++ q_and_a_app/controllers/comment_controller.py
@@ -1,10 +1,8 @@
-from flask import Blueprint, request, render_template, redirect, url_for, current_app, abort, flash
+from flask import Blueprint, request, redirect, url_for, flash, abort
 from main import db
 from models.posts import Post
 from models.comments import Comment
 from schemas.comment_schema import comment_schema, comments_schema
-from schemas.post_schema import posts_schema, post_schema
-import boto3
 from flask_login import login_required, current_user
 
 comments = Blueprint('comments', __name__)
@@ -14,14 +12,12 @@ comments = Blueprint('comments', __name__)
 @login_required
 def create_comment(id):
     text = request.form.get('text')
-    
     if not text:
         flash('Comment cannot be empty.', category='error')
     else:
         post = Post.query.get_or_404(id)
         if post:
             comment = Comment(text = text, commentor_id = current_user.id, post_id = id)
-            
             db.session.add(comment)
             db.session.commit()
         else:
@@ -29,6 +25,7 @@ def create_comment(id):
     return redirect(url_for('posts.get_feed'))
 
 
+# Like a comment
 @comments.route("/feed/<int:id>/like_comment/", methods=["POST"])
 @login_required
 def like_comment(id):
@@ -38,6 +35,7 @@ def like_comment(id):
     return redirect(url_for('users.user_detail'))
 
 
+# Unlike a comment
 @comments.route("/feed/<int:id>/unlike_comment/", methods=["POST"])
 @login_required
 def unlike_comment(id):
@@ -47,12 +45,13 @@ def unlike_comment(id):
     return redirect(url_for('users.user_detail'))
 
 
-
-# Delete a specific post
+# Delete a comment
 @comments.route('/feed/<int:id>/delete_comment/', methods=["POST"])
 @login_required
 def delete_post(id):
     comment = Comment.query.filter_by(comment_id=id).first()
+    if current_user.id != comment.commentor_id:
+        abort(403, "You do not have permission to delete this comment!")
     db.session.delete(comment)
     db.session.commit()
     return redirect(url_for("posts.get_feed"))
